refactor: 컴포넌트는 모두 클래스로 관리되도록 수정
