+++ src/DOMController.ts
@@ -1,6 +1,6 @@
 import { $ } from "./utils/selector";
 import MoviePosterBoard from "./components/MoviePosterBoard/MoviePosterBoard";
-import createHeader from "./components/Header/createHeader";
+import Header from "./components/Header/Header";
 
 class DOMController {
   static moviePosterBoard = new MoviePosterBoard("popular");
@@ -13,7 +13,7 @@ class DOMController {
   }
 
   static renderHeader() {
-    const header = createHeader();
+    const header = new Header().element;
 
     $("body")?.append(header);
   }
