+++ src/components/Button/common/createButton.ts
@@ -1,4 +1,4 @@
-import createElement from "../../utils/createElement";
+import createElement from "../../../utils/createElement";
 
 const createButton = (
   contents: string,
